start_pts int64

+++ probedata.go
@@ -56,7 +56,7 @@ type Program struct {
 	StreamsCount  int         `json:"nb_streams"`
 	PmtPid        int         `json:"pmt_pid"`
 	PcrPid        int         `json:"pcr_pid"`
-	StartPts      int32       `json:"start_pts"`
+	StartPts      int64       `json:"start_pts"`
 	StartTime     float32     `json:"start_time"`
 	Tags          *ProgramTag `json:"tags"`
 	Streams       []*Stream   `json:"streams"`
